relocated SubtractFieldOptions/fixed argparser / test
